requirements.txt
fastapi>=0.100.0
sqlalchemy>=2.0.0
fastapi_pagination>=0.12.6
pydantic>=2.0.0
python-dateutil>=2.8.2

black
isort
pylint
mypy